+++ src/collective/fingerpointing/utils.py
@@ -6,13 +6,19 @@ from zope.globalrequest import getRequest
 def get_request_information():
     """Return logged in user name and remote IP address."""
     request = getRequest()
-    user = api.user.get_current().getMemberId()
+    try:
+        user_id = api.user.get_current().getMemberId()
+    except AttributeError:
+        user_id = '-'  # anonymous user
 
     # honor Cloudflare real client IP address request header if present
     # see: https://support.cloudflare.com/hc/en-us/articles/200170986
     if 'HTTP_CF_CONNECTING_IP' in request.environ:
         ip = request.environ['HTTP_CF_CONNECTING_IP']
+    # Common proxy configuration
+    elif 'HTTP_X_FORWARDED_FOR' in request.environ:
+        ip = request.environ['HTTP_X_FORWARDED_FOR']
     else:
-        ip = request.getClientAddr() or 'None'
+        ip = request.getClientAddr() or 'None'  # return 'None' on tests
 
-    return user, ip  # returns ('test_user_1_', 'None') on tests
+    return user_id, ip